fix: restrict countme worker entrypoint exports to default handler
Move route mappings and SVG helpers to routes.mjs to comply with
Cloudflare Workers entrypoint export requirements while preserving
offline unit testability. Update cloudflare-workers skill.
